sequencer
The fifteen lines of code exercised a concise algorithmic function, sifting through an input text, and extracting and cataloguing letters which matched those used to abbreviate the nucleotides found in deoxyribonucleic acid (DNA)³. The catalogue, in the form of a list, was then adjoined into a single string of text. That string of text, which could be thousands of characters long, could be fed into a genomic database, such as the National Library of Medicine’s Basic Local Alignment Search Tool (BLAST), and compared to a litany of genetic sequences researchers had gleaned from flora and fauna found on our planet.